hwacha: Sign-extend float-point vector register elements
This matches the actual Hwacha RTL.  Although unconditionally filling
the upper bits with ones has the benefit that narrower floating-point
values are treated as NaNs when interpreted as wider types, it also
introduces inconsistences in a mixed-precision environment.

As a generalization of the baseline case with fixed-width elements,
registers of any width must be permitted to hold any data of equal or
lesser precision.  It should be possible to reduce the width of a given
register while preserving the behavior of the original code, assuming
that the change does not cause trunction of higher-precision values.

Consider the following instruction sequence:

	vfmadd.s vv0, ...
	vsd vv0, ...

If vv0 is configured with 64-bit elements, then under the prior model,
the upper 32 bits consist of all ones.  However, if vv0 is configured
with 32-bit elements, then it becomes ambiguous how the vsd instruction
should extend the store data to 64 bits.  This arises from vector
registers being agnostic to whether they hold integer or floating-point
values.  Applying sign-extension uniformly, regardless of datatype,
resolves this ambiguity.